last updates for task 2
i created the reservation and details.
Now i need to see how i can link them properly between one table and another because postman was bugging while i was trying to se if the connection works. so far there are a few tables that work that are liked one to another. but thats for task 3 to figure out what i can do to fix it through postman